style: Restored restaurant image for narrow displays
Based on reviewer input unhid the restaurant image on the
restaurant page also for narrow displays. My understanding
of the quoted rubric in the project specification:
---
The detail page includes a map, hours and reviews in all viewports.
---
...was that hiding the image (= not essential information) on small
displays was preferred. Therefore hiding it was a last-minute
change. It is now back.