feat: Download from server, save to IDB
Removed the old restaurants.json-file. Changed DBHelper DATABASE_URL
to point to the sail server on localhost / port 1337. Modified
fetchRestaurants and fetchRestaurantById to work based on the
(slightly simpler) format returned by the server. Embedded the
idb.js library interface code to dbhelper.js. Created the IDB
calls to default to loading the requested resource from the db.
If not available, the resource is fetched from the server and
saved to the db. Created also an additional keyval objectstore
to the same db the save the status of whether the full set of
restaurants has been fetched. Made minor modifications to
imageUrlForRestaurant and smallImageUrlForRestaurant to comply
with the changes in naming compared to the restaurants.json file.
The previous version also had figure captions for alt text added
by me. These were not available in the sail database, so changed
altTextForRestaurantImage to serve the boring restaurant.name
instead.