feat: Source optimization
Introduced gulp and all the changes, which came together with it.
The original webapp is now in an 'src' directory. Gulp can be
used to concatenate and minify it in two steps: first to a 'dist'
directory, which also has new versions of the *.html and sw.js
files (referencing the concatenated js-files) and then to
project root, which will compress also the *.html and sw.js
files. Also the restaurant-page map is now only loaded on-demand.