feat: User input processing
Favorite restaurant marking buttons (with Aria features) added on the
home page. A pop-up form added to enter new reviews into the database.
If review posting fails, the review is added to an IDB store and
a sync event is scheduled for offline posting of the review. If offline
posting fails, it is re-scheduled. New reviews are also added directly
to the page HTML, so they are immediately visible on the page
independent of connectivity. When posting to server succeeds, the
IDB database is synced with the server for the latest set of reviews.