feat: Page load optimisations
Removed caching of images during service worker install, as it
was caching both image sizes, which caused delay. Used gulp-webp to
compress all images to smaller size and change to webp format. Because
the dbhelper.js script also had to change to load .webp files, changed
also the image directories under src-directory to contain the
webp-format files, so that the same js would be usable also
for debug-runs on source. Moved the original jpg-images to another
folder for storage.